Extract StorageManager from PersistentEngine
- Created `StorageManager` in `relvar-storage/src/storage/manager.rs` to handle `Catalog` and `HeapFile` management.
- Moved physical storage logic and recovery of uncommitted inserts from `PersistentEngine` to `StorageManager`.
- Updated `PersistentEngine` to delegate to `StorageManager`, improving cohesion and separation of concerns.
- Updated `load_relation` in `StorageManager` to take `&self` (immutable reference) for concurrent read access, bypassing the internal cache which is used for writes.
- Fixed a clippy warning regarding collapsible if statements.
- Verified changes with full test suite.
