+++ relvar-storage/src/storage/mod.rs
@@ -63,9 +63,11 @@
 
 pub mod catalog;
 pub mod heap;
+pub mod manager;
 pub mod page;
 
 pub use catalog::{Catalog, CatalogError};
 pub use heap::{HeapError, HeapFile};
+pub use manager::StorageManager;
 // TupleId is now pub(crate) in heap.rs, not exported
 pub use page::{PAGE_SIZE, Page, PageError, PageFile, PageId};
